Allow running with old Servlet API specifications
Our build time dependency to the javax.servlet.http package is through the 3.1
version of Servlet API. But given the limited usage we make of this package, it
is compatible with all versions. So we put 0.0.0 as version for the imported
package in the OSGi manifest, which means "any version".

Other imported packages have version restriction that look reasonable:
- jackson: we use 2.5.2 at build time, so [2.5,3) range is OK
- spring: we use 3.1.2 at build time, so [3.1,4) range looks good;
someone with better knowledge of API compatibility may suggest we use [3.0,4)
if compatible
- javax.portlet: we use 2.0 at build time, and the classes we use were
introduced in 2.0, so [2.0,3] range is OK