#306: extract common HTTP request and response interfaces inside JsonRpcServer
Common logic may be used to handle the HTTP requests.
This is a preparation for using jakarta.servlet.http namespace objects.
There are no much requirements for those interfaces,
only some common methods, which are present in most

Portlet API requests were not changed, because there is a comment,
which assumes that HTTP status code is assigned outside.
Added a TODO comment there with possible solution proposal.


Signed-off-by: cyb3r4nt <104218001+cyb3r4nt@users.noreply.github.com>