Pull properties from org.springframework.remoting.support.RemoteExporter down into AbstractJsonServiceExporter
RemoteExporter is deprecated and is now removed from Spring Framework v6+.
It was used only to provide service and serviceInterface properties
and some related methods.
User defined Java services are registered in the Spring MVC using BeanNameUrlHandlerMapping,
and RemoteExporter provided only some base functionality for accessing service class and interface.
This functionality may be pulled down to AbstractJsonServiceExporter,
and this preserves compatibility in existing applications,
which use service and serviceInterface properties.


Signed-off-by: cyb3r4nt <104218001+cyb3r4nt@users.noreply.github.com>